Game of concentration using JavaScript
Cards are randomly laid face down and two cards are flipped face up at
each turn; where the object of the game is to turn over two matching
cards. It contain two rows of five cards each, for a total of 10 cards.
When
the game concludes, the users are given a score in the form of minutes
(i.e. the number of minutes that it took the user to solve the puzzle).
The JavaScript code is well organized and written, and used XML for
documentation. The user experience of the solution is equivalent to no
matter the screen size, meaning that the user always see two rows of
cards without any horizontal or vertical scrolling.